requirements.txt
requests==2.25.1
python-telegram-bot==13.7
pytz==2021.1
psutil==5.8.0